Add renderDecodedProperty
There are two different uses for when you want a property that returns a
rendering function.

The first, is when the value being passed to the
rendering function is an Elm value. An example of this is in a ListView.
The rendering function renders each element in the list, which is
already an Elm value, so there's no need to encode/decode.

The second, is when the value being passed into the rendering function
is a JavaScript value. An example of this is in NavigationExperimental,
where the value being passed to the renderer comes from React Native,
and needs to be decoded before it can be called in the renderer.

This commit renames the existing renderProperty to
renderDecodedProperty, and adds a new function called renderProperty
which does no decoding.